% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{crr.boot}
\alias{crr.boot}
\alias{crr.boot.ci}
\title{Esperimento bootstrap per design NCRR.}
\usage{
crr.boot(
  data,
  statistic,
  R,
  ran.gen,
  mle,
  retain.data = TRUE,
  ...,
  parallel = FALSE,
  nclus = NA,
  trace = 1,
  seed = NULL
)

crr.boot.ci(
  data,
  statistic,
  R,
  ran.gen,
  mle,
  ...,
  within = FALSE,
  signif = 0.05,
  side = c("lower", "upper", "both"),
  interval,
  grid.len,
  psi.grid = seq(interval[1], interval[2], length.out = grid.len),
  parallel = FALSE,
  nclus = NA,
  trace = 1,
  seed = NULL,
  ignore.attrs = FALSE
)
}
\arguments{
\item{data}{Un oggetto \code{crr.design}. Nel caso di \code{crr.boot.ci}, può essere
anche un oggetto di classe \code{crr.boot} (si vedano i dettagli)}

\item{statistic}{Funzione che calcola la statistica e che prende in input
come primo argomento i dati e come secondo argomento il valore in cui
essere calcolata.}

\item{R}{Numero di iterazioni}

\item{ran.gen}{Funzione che dati in input i dati e la stima puntuale dei
parametri genera un nuovo dataset}

\item{mle}{Stima dei parametri per i dati di partenza}

\item{retain.data}{Se TRUE, tiene in memoria tutti i dataset generati
nell'attributo "data" dell'oggetto finale}

\item{...}{Eventuali ulteriori argomenti passati alla statistica.}

\item{parallel}{Indica se utilizzare il calcolo parallelo}

\item{nclus}{Numero di cluster da usare per il calcolo parallelo}

\item{trace}{Intero non negativo che indica l'ammontare di informazioni di
debug stampate a schermo. Se maggiore di zero, stampa una barra di
avanzamento per la procedura bootstrap.}

\item{seed}{Il seed da usare per riprodurre l'esperimento. Può essere un
singolo intero od un vettore di sei elementi}
}
\value{
\emph{Completare...}
}
\description{
\code{crr.boot} calcola la stima bootstrap parametrica di una statistica e
\code{crr.boot.ci} riporta gli intervalli di confidenza bootstrap.
}
\author{
Marco Bressan
}
