% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crr_parameters.R
\name{crr.mean}
\alias{crr.mean}
\alias{crr.vcov}
\title{Parametri per la verosimiglianza normale}
\usage{
crr.mean(params, design = c(0, 1))

crr.vcov(params, design = c(0, 1))
}
\arguments{
\item{params}{Il vettore dei parametri.}

\item{design}{Un vettore intero indicante il design dello studio in esame.}
}
\value{
Un vettore od una matrice di parametri della verosimiglianza dello
specifico studio.
}
\description{
Funzioni a basso livello per il calcolo dei parametri della distribuzione
marginale (normale) della NCRR
}
\details{
NOTA: le funzioni \verb{crr.vcov.*()} sono state deprecate e non dovrebbero più
essere utilizzate
}
\author{
Marco Bressan
}
