% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crr_parameters.R
\name{crr.transform.par}
\alias{crr.transform.par}
\alias{crr.remove.par}
\alias{crr.join.par}
\title{Trasformazione dei parametri}
\usage{
crr.transform.par(params, which = NULL, inverse = FALSE)

crr.remove.par(params, what)

crr.join.par(params, ..., transform = FALSE)
}
\arguments{
\item{params}{Vettore o lista dei parametri}

\item{which}{Se \code{params} non ha i nomi, indica di quale parametro si tratta
(assumendo che tutto il vettore appartenga ad uno stesso parametro, ad
esempio \code{alpha})}

\item{inverse}{Effettua la trasformazione inversa (da scala trasformata ad
originale)?}

\item{what}{Nomi dei parametri da rimuovere}

\item{...}{Può essere usato al posto di \code{params} per specificare i parametri}
}
\value{
I parametri trasformati, nello stesso formato di \code{params} (lista o
vettore).
}
\description{
Trasforma i parametri rispettandone il formato (lista o vettore)
}
\author{
Marco Bressan
}
