% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_getters.R
\name{match.vcov.type}
\alias{match.vcov.type}
\alias{set.vcov.params}
\alias{match.vcov.fixed}
\title{Utilità per la struttura di covarianza}
\usage{
match.vcov.type(type = c("normal", "achana", "equivar", "simple"))

set.vcov.params(params, np = length(params$beta), type)

match.vcov.fixed(type, value = FALSE, np)
}
\arguments{
\item{type}{Uno dei tipi predefiniti di struttura della matrice di
varianza-covarianza}

\item{params}{Il vettore dei parametri sottoforma di lista}

\item{np}{Il numero di trattamenti nella metanalisi escludendo il \emph{baseline}}

\item{value}{Se FALSE restituisce solo i nomi dei parametri da tenere
fissi, altrimenti anche i valori che essi devono assumere.}
}
\description{
Queste funzioni sono state pensate per facilitare l'utente del pacchetto
nella definizione di esperimenti NCRR con strutture di varianza-covarianza
tra studi più o meno flessibili.
}
\details{
\emph{elencare tutte le matrici presenti}
}
\author{
Marco Bressan
}
