% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{crr.optim}
\alias{crr.optim}
\title{Ottimizzazione di una network crr.}
\usage{
crr.optim(
  design,
  init,
  rs = 1,
  transform = TRUE,
  optim.method = "Nelder-Mead",
  optim.control = list(),
  ...
)
}
\arguments{
\item{design}{Disegno dello studio (creato con l'apposita funzione)}

\item{init}{Valori iniziali in formato lista. I parametri esplicitamente
settati a \code{NULL} sono assunti fissati; quelli mancanti vengono inferiti
automaticamente.}

\item{rs}{Numero di random-start. Se impostato, sovrascrive i valori di
partenza specificati da init}
}
\value{
Risultato dell'ottimizzazione
}
\description{
Funzione per l'ottimizzazione centralizzata
}
\author{
Marco Bressan
}
