% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstar.R
\name{crr.rstar}
\alias{crr.rstar}
\title{Calcolo della statistica di Skovgaard}
\usage{
crr.rstar(
  data,
  thetainit,
  floglik,
  fscore = NULL,
  fpsi,
  psival,
  datagen,
  R = 1000,
  seed = NULL,
  ronly = FALSE,
  parallel = FALSE,
  nclus = NA,
  trace = 1,
  precalc = NULL
)
}
\arguments{
\item{data}{Un oggetto di classe \code{ncrr.design}}

\item{thetainit}{Stima iniziale del vettore dei parametri}

\item{floglik}{Funzione di verosimiglianza, idealmente il prodotto di una
chiamata a \code{get.llik.from.design()}}

\item{fscore}{Funzione score, se disponibile (altrimenti viene usata
l'approssimazione numerica del gradiente contenuta in \code{pracma})}

\item{fpsi}{Funzione scalare che estrapola il valore del parametro
d'interesse}

\item{psival}{Valore del parametro d'interesse sotto l'ipotesi nulla}

\item{datagen}{Funzione che genera il dataset a partire dalla stima
corrente dei parametri}

\item{R}{Numero di iterazioni per il calcolo della statistica di Skovgaard
tramite Montecarlo}

\item{seed}{Seed per il metodo Montecarlo}

\item{ronly}{Se TRUE, non calcola la statistica di Skovgaard}

\item{parallel}{Calcolo parallelo per la statistica di Skovgaard}

\item{nclus}{Numero di cluster per il calcolo parallelo}

\item{trace}{Intero >= 0}

\item{precalc}{Lista di valori precalcolati (SPERIMENTALE!!)}
}
\value{
\emph{Da completare...}
}
\description{
Calcolo della statistica di Skovgaard
}
\author{
Marco Bressan
}
