% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crr_parameters.R
\name{crr.split.par}
\alias{crr.split.par}
\title{Separa i parametri di una NCRR}
\usage{
crr.split.par(
  params,
  np = NA,
  transform = FALSE,
  fixed = NULL,
  parlen = attr(fixed, "parlen")
)
}
\arguments{
\item{params}{vettore numerico con i parametri in formato "lungo"}

\item{np}{lunghezza degli alpha - 1. Settando a \code{NA} si tenta di calcolarlo
sulla base dei nomi del vettore dei parametri.}

\item{transform}{ritorna i parametri trasformati in scala logaritmica?}

\item{fixed}{elenco di parametri da escludere. È raccomandato che si
utilizzi l'output di \code{match.vcov.fixed()}}

\item{parlen}{Vettore di interi avente i nomi corrispondenti ai parametri e
la lunghezza degli stessi. Se non specificato, legge l'attributo 'parlen'
da \code{fixed}. ATTENZIONE: per l'utilizzo standard, si consiglia di non
settare esplicitamente questo parametro, ma affidarsi a \code{fixed}.}
}
\description{
Lo scopo di questa funzione è fornire un tramite tra la rappresentazione
dei parametri sottoforma di vettore voluta da \code{optim} e una più
"user-friendly" in cui i parametri sono separati
}
\author{
Marco Bressan
}
