% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{get.matrix.from.design}
\alias{get.matrix.from.design}
\title{Ottieni matrici dal design}
\usage{
get.matrix.from.design(object, what = c("theta", "gamma"))
}
\arguments{
\item{object}{design di uno studio ncrr}

\item{what}{al momento solo "theta" o "gamma"}
}
\value{
matrice di interesse (quella dei gamma o dei theta) in un
formato che sia compatibile con la notazione usata in maxima nel calcolo
delle derivate esplicite.
}
\description{
L'uso della presente è confinato alle funzioni che calcolano i
parametri esplicitamente, contenute nella dir \verb{data-raw/}
}
\author{
Marco Bressan
}
