% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncrr.design_class.R
\name{getInitial.ncrr.design}
\alias{getInitial.ncrr.design}
\title{Parametri iniziali per l'ottimizzatore della verosimiglianza NCRR}
\usage{
\method{getInitial}{ncrr.design}(
  object,
  data,
  ...,
  eps = 1e-10,
  seed = if (rep > 1) NA,
  rep = 1,
  transform = TRUE,
  fixed = NULL,
  vcov.type = attr(object, "vcov.type") \%||\% "normal"
)
}
\arguments{
\item{object}{Oggetto di tipo \code{ncrr.design}}

\item{data}{Non usato}

\item{...}{Non usati}

\item{eps}{Tolleranza numerica per lo 0}

\item{seed}{Seme per la generazione di un punto di partenza casuale. Se si
passa un singolo valore, esso viene usato per la generazione di un seed
diverso per ogni parametro (con NA tale generazione è casuale). Può
essere un vettore con nomi dei parametri \code{c(alpha, beta, mu0, sigma20, rho, sigma2)}: per specificare quali NON inizializzare casualmente,
mettere NA. Impostando \code{seed} a \code{NULL}, la funzione ritorna dei valori di
default, deterministici.}

\item{rep}{numero di random start. Ignorato se il \code{seed} è \code{NULL}}

\item{transform}{applica le trasformazioni ai parametri?}

\item{fixed}{elenco di nomi di parametri fissati (cioè esclusi dal
vettore/matrice)}

\item{vcov.type}{preimpostazioni sui parametri fixed}
}
\value{
un vettore o una matrice con i punti di partenza casuali (se rep>1)
}
\description{
Parametri iniziali per l'ottimizzatore della verosimiglianza NCRR
}
