% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncrr.design_class.R
\name{labbe.plot}
\alias{labbe.plot}
\title{L'Abbé Plot}
\usage{
labbe.plot(
  object,
  cex.factr = 3,
  abline.col = "gray30",
  legend.pos = "bottomright",
  xylims = NULL,
  xlab = "Perc. of events in control group",
  ylab = "Perc. of events in treatment group"
)
}
\arguments{
\item{object}{Un oggetto contenente i dati necessari per il grafico.}

\item{legend.pos}{Posizione della legenda (default: "bottomright"). Imposta a NULL per non visualizzare la legenda.}

\item{point.size.fattore}{Fattore di scala per la dimensione dei punti (default: 3).}

\item{baseline.colore}{Colore della linea di baseline (default: "gray30").}
}
\description{
Crea un grafico di L'Abbé plot a partire da un oggetto
di classe \code{ncrr.design}
}
\examples{
object <- ncrr.design(smoking)
labbe.plot(object)

}
