% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{get.llik.from.design}
\alias{get.llik.from.design}
\alias{get.llik.from.design2}
\title{Funzione per la definizione della verosimiglianza per NCRR.}
\usage{
get.llik.from.design(
  object,
  transform = TRUE,
  echo = 0,
  vcov.type = attr(object, "vcov.type"),
  use.data = FALSE,
  stop.on.fail = TRUE
)

get.llik.from.design2(
  object,
  transform = TRUE,
  echo = 0,
  vcov.type = attr(object, "vcov.type"),
  use.data = FALSE,
  stop.on.fail = TRUE
)
}
\arguments{
\item{object}{oggetto che definisce il design}

\item{transform}{applica trasformazioni ai parametri di
varianza/correlazione}

\item{echo}{regola il livello delle stampe di debug}

\item{vcov.type}{struttura della matrice di varianza-covarianza}

\item{use.data}{se \code{TRUE}, la funzione risultante prenderà in input il dataset
completo anzichè le singole componenti. Utile per l'utilizzo in combinazione
con la libreria \code{boot} o \code{likelihoodAsy}.}

\item{stop.on.fail}{La funzione dovrebbe restituire un errore se una
componente della verosimiglianza risulta NA? Se l'argomento è impostato a
FALSE oppure NA, la funzione restituirà invece -Inf, nel primo caso con un
avvertimento.}
}
\value{
una funzione del vettore dei parametri (in tal senso \code{llik1} ne
costituisce una versione semplificata), con la possibilità di fissare gli
stessi (opzione \code{fixed = list(...)})
}
\description{
A partire da un \emph{design}, crea un oggetto funzione che può essere
passato ad \code{optim}.
}
\details{
\code{get.llik.from.design2} è una specializzazione per design contenenti solo studi
con due trattamenti.
}
\author{
Marco Bressan
}
