% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{llik1}
\alias{llik1}
\title{Verosimiglianza NCRR}
\usage{
llik1(params, y, Gamma, design = c(0, 1))
}
\arguments{
\item{params}{vettore dei parametri contenuti in un unico vettore; in ordine: alfa_0j, beta_0j, mu0, sigma^2_0, rho, sigma^2_{0j}.}

\item{y}{vettore dei vartheta}

\item{Gamma}{matrice Gamma}

\item{design}{vettore che specifica il design (attualmente utile solo la sua dimensione)}
}
\value{
valore della log-verosimiglianza
}
\description{
Verosimiglianza per un singolo studio per NCRR \strong{con baseline}.
--VERSIONE PRELIMINARE--
}
\author{
Marco Bressan
}
