% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncrr.design_class.R
\name{plot.ncrr.design}
\alias{plot.ncrr.design}
\title{Rappresentazione grafica di una Network Meta-Analisi}
\usage{
\method{plot}{ncrr.design}(
  object,
  vertex.size.factr = 50,
  edge.width.factr = 5,
  pos.offset = 0,
  smart.layout.seed = NA,
  ...
)
}
\arguments{
\item{object}{Un oggetto di tipo ncrr.design.}

\item{vertex.size.factr}{Fattore di scala per la dimensione dei vertici (default: 50).}

\item{edge.width.factr}{Fattore di scala per la larghezza degli archi (default: 1).}

\item{pos.offset}{Angolo in radianti di traslazione della disposizione dei nodi rispetto
all'estremo superiore della circonferenza unitaria (default: 0). Ignorato se \code{smart.layout.seed}
è diverso da \code{NA}.}

\item{smart.layout.seed}{Imposta il seed per la generazione di un layout casuale
"intelligente", ovvero con meno intersezioni di archi possibile. Se \code{NA}, disabilita
la funzionalità e i nodi verranno disposti in cerchio.}

\item{...}{Parametri aggiuntivi passati a \code{plot.igraph()}}
}
\description{
Disegna un grafo a partire da un oggetto \code{ncrr.design}.
}
\examples{
object <- ncrr.design(smoking)
plot(object)

}
