% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate.ncrr.design}
\alias{simulate.ncrr.design}
\title{Funzione per simulare da un design. Può essere usato in combinazione con
\code{random.design()} per svolgere studi di simulazione approfonditi.}
\usage{
\method{simulate}{ncrr.design}(
  object,
  nsim = 1,
  seed = rpois(1, 1e+05),
  vcov.type = attr(object, "vcov.type"),
  params,
  ...
)
}
\arguments{
\item{object}{Un oggetto di classe \code{ncrr.design}}

\item{nsim}{numero di simulazioni}

\item{seed}{Seme casuale per riproducibilità}

\item{vcov.type}{Tipo di matrice di varianza-covarianza da usare}

\item{params}{Parametri previsti per la specifica tipologia di matrice (un
sottoinsieme di alpha, beta, mu0, sigma20, rho, sigma2)}

\item{...}{Variante per specificare i parametri come in \code{params}}
}
\value{
Una lista di oggetti \code{ncrr.design} con i dati simulati.
}
\description{
Funzione per simulare da un design. Può essere usato in combinazione con
\code{random.design()} per svolgere studi di simulazione approfonditi.
}
