% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subst.params}
\alias{subst.params}
\title{Sostituisci parametri}
\usage{
subst.params(params, subst)
}
\arguments{
\item{params}{Lista o vettore dei parametri}

\item{subst}{Lista o vettore delle sostituzioni.
La sostituzione avviene in base al nome, che dunque è obbligatorio per
ogni parametro.}
}
\value{
\code{params} opportunamente modificato
}
\description{
Effettua una sostituzione nella lista dei parametri di una NCRR
}
\author{
Marco Bressan
}
