% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crr_parameters.R
\name{crr.split.par}
\alias{crr.split.par}
\title{separa i parametri di una ncrr}
\usage{
crr.split.par(
  params,
  np,
  transform = FALSE,
  fixed = NULL,
  parlen = attr(fixed, "parlen")
)
}
\arguments{
\item{params}{vettore numerico con i parametri in formato "lungo"}

\item{np}{lunghezza degli alpha - 1}

\item{transform}{ritorna i parametri trasformati in scala logaritmica?}

\item{fixed}{elenco di parametri da escludere}
}
\value{
una lista con i parametri separati
}
\description{
lo scopo di questa funzione è fornire un tramite tra la rappresentazione
dei parametri sottoforma di vettore voluta da \code{optim} e una più
"user-friendly" in cui i parametri sono separati
}
\author{
Marco Bressan
}
