% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{get.llik.from.design}
\alias{get.llik.from.design}
\title{Funzione per il calcolo della verosimiglianza per NCRR.}
\usage{
get.llik.from.design(object)
}
\arguments{
\item{object}{oggetto che definisce il design}
}
\value{
una funzione con parametro come \code{llik1}
}
\description{
Funzione per il calcolo della verosimiglianza per NCRR.
}
