% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{getInitial.ncrr.design}
\alias{getInitial.ncrr.design}
\title{Parametri iniziali per l'ottimizzatore della verosimiglianza NCRR}
\usage{
\method{getInitial}{ncrr.design}(object, data, ..., eps = 1e-10)
}
\arguments{
\item{object}{Oggetto di tipo \code{ncrr.design}}

\item{data}{Non usato}

\item{...}{Non usati}

\item{eps}{Tolleranza numerica per lo 0}
}
\description{
Parametri iniziali per l'ottimizzatore della verosimiglianza NCRR
}
