% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{labbe.plot}
\alias{labbe.plot}
\title{L'Abbé Plot}
\usage{
labbe.plot(
  object,
  point.size.fattore = 3,
  baseline.colore = "gray30",
  legend.pos = "bottomright"
)
}
\arguments{
\item{object}{Un oggetto contenente i dati necessari per il grafico.}

\item{point.size.fattore}{Fattore di scala per la dimensione dei punti (default: 3).}

\item{baseline.colore}{Colore della linea di baseline (default: "gray30").}

\item{legend.pos}{Posizione della legenda (default: "bottomright"). Imposta a NULL per non visualizzare la legenda.}
}
\description{
Crea un grafico di L'Abbé plot a partire da un oggetto
di classe \code{ncrr.design}
}
\examples{
# labbe.plot(object)

}
