% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{ncrr.design}
\alias{ncrr.design}
\title{Specifica il disegno della meta-analisi}
\usage{
ncrr.design(x, baseline)
}
\arguments{
\item{x}{un dataframe}

\item{baseline}{livello baseline (di default il primo)}
}
\value{
lista con i dettagli dell'esperimento
}
\description{
\strong{NOTA:} al momento bisogna assicurarsi che il dataframe sia ordinato per studio
e per trattamento, con il baseline in prima posizione
}
