% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncrr.design_class.R
\name{ncrr.design}
\alias{ncrr.design}
\title{Specifica il disegno della meta-analisi}
\usage{
ncrr.design(x, vcov.type = NULL)
}
\arguments{
\item{x}{un dataframe. Se contenente le colonne $m_{ik}$ e $s_{ik}$, queste
vengono interpretate come medie e dev.std dello studio, che viene quindi
assunto ad outcome normale. In questo caso,
$\hat\Gamma=\text{diag}(s_{ik}^2)$ \code{x$treatment} dev'essere un fattore: il
primo livello sarà il \emph{baseline}.}
}
\value{
lista con i dettagli dell'esperimento
}
\description{
Specifica il disegno della meta-analisi
}
