% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{plot.ncrr.design}
\alias{plot.ncrr.design}
\title{Rappresentazione grafica di una Network Meta-Analisi}
\usage{
\method{plot}{ncrr.design}(
  object,
  vertex.size.factr = 50,
  edge.width.factr = 1,
  vertex.color = "lightblue",
  edge.color = "gray"
)
}
\arguments{
\item{object}{Un oggetto di tipo ncrr.design.}

\item{vertex.size.factr}{Fattore di scala per la dimensione dei vertici (default: 50).}

\item{edge.width.factr}{Fattore di scala per la larghezza degli archi (default: 1).}

\item{vertex.color}{Colore dei vertici (default: "lightblue").}

\item{edge.color}{Colore degli archi (default: "gray").}
}
\description{
Disegna un grafo a partire da un oggetto \code{ncrr.design}.
}
\examples{
object <- ncrr.design(smoking)
plot.ncrr.design(object)

}
