% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rp_stat.R
\name{rp.stat}
\alias{rp.stat}
\title{Calcolo della statistica log-RV}
\usage{
rp.stat(
  dati.gen,
  psi0,
  init,
  param = match("beta5", names(init)),
  theta.hat = NULL,
  J = NULL,
  ...,
  exact = NA,
  par.only = FALSE,
  export = c("theta.hat", "l.hat", "J")
)
}
\arguments{
\item{dati.gen}{Oggetto di classe ncrr.design}

\item{psi0}{Valore del parametro di interesse $\psi$ sotto l'ipotesi nulla.}

\item{init}{Vettore di valori iniziali per l'ottimizzazione dei parametri.}

\item{param}{Indice del parametro di interesse $\psi$. Di default si
considera il parametro "beta5".}

\item{theta.hat}{(Opzionale) Punto di massima della verosimiglianza
precalcolato. Se NULL viene stimata internamente.}

\item{J}{(Opzionale) Matrice hessiana della funzione di log-verosimiglianza
precalcolata. Se NULL viene stimata internamente.}

\item{...}{Argomenti aggiuntivi passati alla funzione di verosimiglianza.}

\item{exact}{Se TRUE, usa ottimizzazione vincolata esatta.}

\item{par.only}{Se TRUE, restituisce solo i parametri ottimizzati sotto
vincolo}

\item{export}{Nomi degli attributi da esportare: di default, \code{theta.hat}
(stima libera dei parametri), \code{l.hat} (verosimiglianza nel punto di
massimo) e \code{J} (matrice hessiana)}
}
\value{
La statistica $r_p$ come oggetto numerico con gli attributi
specificati in \code{export}. Se \code{par.only = TRUE}, restituisce il vettore dei
parametri ottimizzati sotto vincolo.
}
\description{
Questa funzione calcola la statistica radice con segno ($r_p$), basata
sul log-rapporto di verosimiglianza, utile nell'inferenza parametrica
condizionale e nell'analisi della significatività di parametri di interesse.
}
